requirements.txt
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
dnspython==2.6.1
fast-pagerank==1.0.0
fonttools==4.49.0
kiwisolver==1.4.5
matplotlib==3.8.3
numpy==1.26.4
packaging==24.0
pillow==10.2.0
pymongo==4.6.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
scipy==1.12.0
six==1.16.0
